[FLINK-16123] [kafka] Refactor json spec parsing for reuseability
